Remove sockfd global
sockfd is now a local variable in run_bot() which is passed to the networking
functions as needed.

recv_msg() has been changed from int recv_msg(struct irc_message **message)
to struct irc_message *recv_msg(int). Instead of the return value meaning bytes
read, or <0 for error, it now returns a NULL pointer to mean error, or a
pointer to the irc_message it allocates via create_message if successful.
The bytes read are no longer required by the loop in run_bot().

process_message() now also takes sockfd as a parameter. This change is so that
it may call send_msg() with the sockfd, it does not use it directly.

send_msg() has had sockfd added as a parameter.

run_bot() has also recieved some minor style tweaks.