Fix bug in process message function
Forced the integer variable that keeps track of the number of messages
in a response to reset to zero on running each plugin.

Previously, it was initialized to zero when defined, and never changed
afterwards. This caused a scenario for a plugin to have one response,
and another plugin to be tested and have zero responses, but if the
plugin did not explicitly set the response integer pointer to zero
process_message would see it as having 1 response, try to dereference a
random pointer pointing to who knows where, and crash.