Change plugins to act on params `commands'
Instead of acting on an irc message's command, operate on a custom
command that is extracted out of the params of a PRIVMSG command.

If the PRIVMSG params begins with the defined command char, extract the
following token from the string and test it against plugin defined
commands to see if any will accept it.

This will allow for a documentation system to easily be built on a per
command basis in the future.